requirements.txt
pyrogram==2.0.106
tgcrypto==1.2.5
aiosqlite==0.20.0
pytest==8.3.5
pytest-asyncio==0.24.0
python-dotenv==1.0.1
